moved pydantic schemas to separate file. updated API
